+++ sdk/tables/azure-data-tables/src/main/java/com/azure/data/tables/TableAsyncClient.java
@@ -5,6 +5,7 @@ package com.azure.data.tables;
 import com.azure.core.annotation.ReturnType;
 import com.azure.core.annotation.ServiceClient;
 import com.azure.core.annotation.ServiceMethod;
+import com.azure.core.credential.AzureNamedKeyCredential;
 import com.azure.core.http.HttpHeaders;
 import com.azure.core.http.HttpPipeline;
 import com.azure.core.http.HttpRequest;
@@ -21,15 +22,12 @@ import com.azure.core.util.logging.ClientLogger;
 import com.azure.core.util.serializer.SerializerAdapter;
 import com.azure.data.tables.implementation.AzureTableImpl;
 import com.azure.data.tables.implementation.AzureTableImplBuilder;
-import com.azure.data.tables.implementation.TransactionalBatchImpl;
 import com.azure.data.tables.implementation.ModelHelper;
+import com.azure.data.tables.implementation.TableSasUtils;
+import com.azure.data.tables.implementation.TableServiceSasGenerator;
 import com.azure.data.tables.implementation.TableUtils;
+import com.azure.data.tables.implementation.TransactionalBatchImpl;
 import com.azure.data.tables.implementation.models.AccessPolicy;
-import com.azure.data.tables.implementation.models.TransactionalBatchChangeSet;
-import com.azure.data.tables.implementation.models.TransactionalBatchAction;
-import com.azure.data.tables.implementation.models.TransactionalBatchRequestBody;
-import com.azure.data.tables.implementation.models.TransactionalBatchSubRequest;
-import com.azure.data.tables.implementation.models.TransactionalBatchResponse;
 import com.azure.data.tables.implementation.models.OdataMetadataFormat;
 import com.azure.data.tables.implementation.models.QueryOptions;
 import com.azure.data.tables.implementation.models.ResponseFormat;
@@ -38,7 +36,11 @@ import com.azure.data.tables.implementation.models.TableEntityQueryResponse;
 import com.azure.data.tables.implementation.models.TableProperties;
 import com.azure.data.tables.implementation.models.TableResponseProperties;
 import com.azure.data.tables.implementation.models.TableServiceError;
-import com.azure.data.tables.models.TableTransactionActionResponse;
+import com.azure.data.tables.implementation.models.TransactionalBatchAction;
+import com.azure.data.tables.implementation.models.TransactionalBatchChangeSet;
+import com.azure.data.tables.implementation.models.TransactionalBatchRequestBody;
+import com.azure.data.tables.implementation.models.TransactionalBatchResponse;
+import com.azure.data.tables.implementation.models.TransactionalBatchSubRequest;
 import com.azure.data.tables.models.ListEntitiesOptions;
 import com.azure.data.tables.models.TableAccessPolicy;
 import com.azure.data.tables.models.TableEntity;
@@ -47,8 +49,10 @@ import com.azure.data.tables.models.TableItem;
 import com.azure.data.tables.models.TableServiceException;
 import com.azure.data.tables.models.TableSignedIdentifier;
 import com.azure.data.tables.models.TableTransactionAction;
+import com.azure.data.tables.models.TableTransactionActionResponse;
 import com.azure.data.tables.models.TableTransactionFailedException;
 import com.azure.data.tables.models.TableTransactionResult;
+import com.azure.data.tables.sas.TableServiceSasSignatureValues;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
@@ -193,6 +197,36 @@ public final class TableAsyncClient {
      */
     public TableServiceVersion getServiceVersion() {
         return TableServiceVersion.fromString(tablesImplementation.getVersion());
+    }
+
+    /**
+     * Generates a service SAS for the table using the specified {@link TableServiceSasSignatureValues}.
+     *
+     * <p>Note : The client must be authenticated via {@link AzureNamedKeyCredential}
+     * <p>See {@link TableServiceSasSignatureValues} for more information on how to construct a service SAS.</p>
+     *
+     * @param tableServiceSasSignatureValues {@link TableServiceSasSignatureValues}
+     *
+     * @return A {@code String} representing the SAS query parameters.
+     */
+    public String generateSas(TableServiceSasSignatureValues tableServiceSasSignatureValues) {
+        return generateSas(tableServiceSasSignatureValues, Context.NONE);
+    }
+
+    /**
+     * Generates a service SAS for the table using the specified {@link TableServiceSasSignatureValues}.
+     *
+     * <p>Note : The client must be authenticated via {@link AzureNamedKeyCredential}
+     * <p>See {@link TableServiceSasSignatureValues} for more information on how to construct a service SAS.</p>
+     *
+     * @param tableServiceSasSignatureValues {@link TableServiceSasSignatureValues}
+     * @param context Additional context that is passed through the code when generating a SAS.
+     *
+     * @return A {@code String} representing the SAS query parameters.
+     */
+    public String generateSas(TableServiceSasSignatureValues tableServiceSasSignatureValues, Context context) {
+        return new TableServiceSasGenerator(tableServiceSasSignatureValues, getTableName())
+            .generateSas(TableSasUtils.extractNamedKeyCredential(getHttpPipeline()), context);
     }
 
     /**