+++ sdk/tables/azure-data-tables/src/main/java/com/azure/data/tables/TableServiceClientBuilder.java
@@ -177,7 +177,8 @@ public final class TableServiceClientBuilder {
     }
 
     /**
-     * Sets the SAS token used to authorize requests sent to the service.
+     * Sets the SAS token used to authorize requests sent to the service. Setting this value will undo other
+     * configuration made with {@code credential(AzureSasCredential)} or {@code credential(AzureNamedKeyCredential)}.
      *
      * @param sasToken The SAS token to use for authenticating requests.
      *
@@ -196,12 +197,14 @@ public final class TableServiceClientBuilder {
 
         this.sasToken = sasToken;
         this.azureNamedKeyCredential = null;
+        this.azureSasCredential = null;
 
         return this;
     }
 
     /**
-     * Sets the {@link AzureSasCredential} used to authorize requests sent to the service.
+     * Sets the {@link AzureSasCredential} used to authorize requests sent to the service. Setting this value will undo
+     * other configuration made with {@code credential(AzureNamedKeyCredential)} or {@code sasToken(String)}.
      *
      * @param credential {@link AzureSasCredential} used to authorize requests sent to the service.
      *
@@ -215,12 +218,15 @@ public final class TableServiceClientBuilder {
         }
 
         this.azureSasCredential = credential;
+        this.azureNamedKeyCredential = null;
+        this.sasToken = null;
 
         return this;
     }
 
     /**
-     * Sets the {@link AzureNamedKeyCredential} used to authorize requests sent to the service.
+     * Sets the {@link AzureNamedKeyCredential} used to authorize requests sent to the service. Setting this will
+     * override any other configuration made with {@code credential(AzureSasCredential)} or {@code sasToken(String)}.
      *
      * @param credential {@link AzureNamedKeyCredential} used to authorize requests sent to the service.
      *
@@ -234,6 +240,7 @@ public final class TableServiceClientBuilder {
         }
 
         this.azureNamedKeyCredential = credential;
+        this.azureSasCredential = null;
         this.sasToken = null;
 
         return this;