+++ sdk/tables/azure-data-tables/src/main/java/com/azure/data/tables/implementation/TableUtils.java
@@ -8,6 +8,7 @@ import com.azure.core.http.rest.PagedFlux;
 import com.azure.core.http.rest.Response;
 import com.azure.core.http.rest.SimpleResponse;
 import com.azure.core.util.CoreUtils;
+import com.azure.core.util.UrlBuilder;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.data.tables.implementation.models.TableServiceErrorException;
 import com.azure.data.tables.implementation.models.TableServiceErrorOdataError;
@@ -17,15 +18,15 @@ import com.azure.data.tables.models.TableServiceException;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-import javax.crypto.Mac;
-import javax.crypto.spec.SecretKeySpec;
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
-import java.nio.charset.StandardCharsets;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
+import java.net.URLEncoder;
 import java.time.Duration;
-import java.util.Base64;
+import java.time.LocalDateTime;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
+import java.time.format.DateTimeFormatter;
+import java.util.Comparator;
 import java.util.Locale;
 import java.util.Map;
 import java.util.TreeMap;
@@ -37,6 +38,24 @@ import static com.azure.core.util.FluxUtil.monoError;
  */
 public final class TableUtils {
     private static final String UTF8_CHARSET = "UTF-8";
+    private static final String INVALID_DATE_STRING = "Invalid Date String: %s.";
+    /**
+     * Stores a reference to the date/time pattern with the greatest precision Java.util.Date is capable of expressing.
+     */
+    private static final String MAX_PRECISION_PATTERN = "yyyy-MM-dd'T'HH:mm:ss.SSS";
+    /**
+     * Stores a reference to the ISO8601 date/time pattern.
+     */
+    private static final String ISO8601_PATTERN = "yyyy-MM-dd'T'HH:mm:ss'Z'";
+    /**
+     * Stores a reference to the ISO8601 date/time pattern.
+     */
+    private static final String ISO8601_PATTERN_NO_SECONDS = "yyyy-MM-dd'T'HH:mm'Z'";
+    /**
+     * The length of a datestring that matches the MAX_PRECISION_PATTERN.
+     */
+    private static final int MAX_PRECISION_DATESTRING_LENGTH = MAX_PRECISION_PATTERN.replaceAll("'", "")
+        .length();
 
     private TableUtils() {
         throw new UnsupportedOperationException("Cannot instantiate TablesUtils");
@@ -178,30 +197,6 @@ public final class TableUtils {
     }
 
     /**
-     * Computes a signature for the specified string using the HMAC-SHA256 algorithm.
-     *
-     * @param base64Key Base64 encoded key used to sign the string
-     * @param stringToSign UTF-8 encoded string to sign
-     *
-     * @return the HMAC-SHA256 encoded signature
-     *
-     * @throws RuntimeException If the HMAC-SHA256 algorithm isn't support, if the key isn't a valid Base64 encoded
-     * string, or the UTF-8 charset isn't supported.
-     */
-    public static String computeHMac256(final String base64Key, final String stringToSign) {
-        try {
-            byte[] key = Base64.getDecoder().decode(base64Key);
-            Mac hmacSHA256 = Mac.getInstance("HmacSHA256");
-            hmacSHA256.init(new SecretKeySpec(key, "HmacSHA256"));
-            byte[] utf8Bytes = stringToSign.getBytes(StandardCharsets.UTF_8);
-
-            return Base64.getEncoder().encodeToString(hmacSHA256.doFinal(utf8Bytes));
-        } catch (NoSuchAlgorithmException | InvalidKeyException ex) {
-            throw new RuntimeException(ex);
-        }
-    }
-
-    /**
      * Parses the query string into a key-value pair map that maintains key, query parameter key, order. The value is
      * stored as a parsed array (ex. key=[val1, val2, val3] instead of key=val1,val2,val3).
      *
@@ -293,5 +288,160 @@ public final class TableUtils {
         } catch (UnsupportedEncodingException ex) {
             throw new RuntimeException(ex);
         }
+    }
+
+    /**
+     * Performs a safe encoding of the specified string, taking care to insert %20 for each space character instead of
+     * inserting the {@code +} character.
+     *
+     * @param stringToEncode String value to encode
+     * @return the encoded string value
+     * @throws RuntimeException If the UTF-8 charset isn't supported
+     */
+    public static String urlEncode(final String stringToEncode) {
+        if (stringToEncode == null) {
+            return null;
+        }
+
+        if (stringToEncode.length() == 0) {
+            return "";
+        }
+
+        if (stringToEncode.contains(" ")) {
+            StringBuilder outBuilder = new StringBuilder();
+
+            int startDex = 0;
+            for (int m = 0; m < stringToEncode.length(); m++) {
+                if (stringToEncode.charAt(m) == ' ') {
+                    if (m > startDex) {
+                        outBuilder.append(encode(stringToEncode.substring(startDex, m)));
+                    }
+
+                    outBuilder.append("%20");
+                    startDex = m + 1;
+                }
+            }
+
+            if (startDex != stringToEncode.length()) {
+                outBuilder.append(encode(stringToEncode.substring(startDex)));
+            }
+
+            return outBuilder.toString();
+        } else {
+            return encode(stringToEncode);
+        }
+    }
+
+    /*
+     * Helper method to reduce duplicate calls of URLEncoder.encode
+     */
+    private static String encode(final String stringToEncode) {
+        try {
+            return URLEncoder.encode(stringToEncode, UTF8_CHARSET);
+        } catch (UnsupportedEncodingException ex) {
+            throw new RuntimeException(ex);
+        }
+    }
+
+    /**
+     * Performs a safe encoding of a url string, only encoding the path.
+     *
+     * @param url The url to encode.
+     * @return The encoded url.
+     */
+    public static String encodeUrlPath(String url) {
+        /* Deconstruct the URL and reconstruct it making sure the path is encoded. */
+        UrlBuilder builder = UrlBuilder.parse(url);
+        String path = builder.getPath();
+        if (path.startsWith("/")) {
+            path = path.substring(1);
+        }
+        path = urlEncode(urlDecode(path));
+        builder.setPath(path);
+        return builder.toString();
+    }
+
+    /**
+     * Given a String representing a date in a form of the ISO8601 pattern, generates a Date representing it with up to
+     * millisecond precision.
+     *
+     * @param dateString the {@code String} to be interpreted as a <code>Date</code>
+     * @return the corresponding <code>Date</code> object
+     * @throws IllegalArgumentException If {@code dateString} doesn't match an ISO8601 pattern
+     */
+    public static OffsetDateTime parseDate(String dateString) {
+        String pattern = MAX_PRECISION_PATTERN;
+        switch (dateString.length()) {
+            case 28: // "yyyy-MM-dd'T'HH:mm:ss.SSSSSSS'Z'"-> [2012-01-04T23:21:59.1234567Z] length = 28
+            case 27: // "yyyy-MM-dd'T'HH:mm:ss.SSSSSS'Z'"-> [2012-01-04T23:21:59.123456Z] length = 27
+            case 26: // "yyyy-MM-dd'T'HH:mm:ss.SSSSS'Z'"-> [2012-01-04T23:21:59.12345Z] length = 26
+            case 25: // "yyyy-MM-dd'T'HH:mm:ss.SSSS'Z'"-> [2012-01-04T23:21:59.1234Z] length = 25
+            case 24: // "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"-> [2012-01-04T23:21:59.123Z] length = 24
+                dateString = dateString.substring(0, MAX_PRECISION_DATESTRING_LENGTH);
+                break;
+            case 23: // "yyyy-MM-dd'T'HH:mm:ss.SS'Z'"-> [2012-01-04T23:21:59.12Z] length = 23
+                // SS is assumed to be milliseconds, so a trailing 0 is necessary
+                dateString = dateString.replace("Z", "0");
+                break;
+            case 22: // "yyyy-MM-dd'T'HH:mm:ss.S'Z'"-> [2012-01-04T23:21:59.1Z] length = 22
+                // S is assumed to be milliseconds, so trailing 0's are necessary
+                dateString = dateString.replace("Z", "00");
+                break;
+            case 20: // "yyyy-MM-dd'T'HH:mm:ss'Z'"-> [2012-01-04T23:21:59Z] length = 20
+                pattern = ISO8601_PATTERN;
+                break;
+            case 17: // "yyyy-MM-dd'T'HH:mm'Z'"-> [2012-01-04T23:21Z] length = 17
+                pattern = ISO8601_PATTERN_NO_SECONDS;
+                break;
+            default:
+                throw new IllegalArgumentException(String.format(Locale.ROOT, INVALID_DATE_STRING, dateString));
+        }
+
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern(pattern, Locale.ROOT);
+        return LocalDateTime.parse(dateString, formatter).atZone(ZoneOffset.UTC).toOffsetDateTime();
+    }
+
+    /**
+     * Parses a query string into a one to many TreeMap.
+     *
+     * @param queryParams The string of query params to parse.
+     * @return A {@code HashMap<String, String[]>} of the key values.
+     */
+    public static Map<String, String[]> parseQueryString(String queryParams) {
+        final TreeMap<String, String[]> retVals = new TreeMap<>(Comparator.naturalOrder());
+
+        if (CoreUtils.isNullOrEmpty(queryParams)) {
+            return retVals;
+        }
+
+        // split name value pairs by splitting on the '&' character
+        final String[] valuePairs = queryParams.split("&");
+
+        // for each field value pair parse into appropriate map entries
+        for (String valuePair : valuePairs) {
+            // Getting key and value for a single query parameter
+            final int equalDex = valuePair.indexOf("=");
+            String key = urlDecode(valuePair.substring(0, equalDex)).toLowerCase(Locale.ROOT);
+            String value = urlDecode(valuePair.substring(equalDex + 1));
+
+            // add to map
+            String[] keyValues = retVals.get(key);
+
+            // check if map already contains key
+            if (keyValues == null) {
+                // map does not contain this key
+                keyValues = new String[]{value};
+            } else {
+                // map contains this key already so append
+                final String[] newValues = new String[keyValues.length + 1];
+                System.arraycopy(keyValues, 0, newValues, 0, keyValues.length);
+
+                newValues[newValues.length - 1] = value;
+                keyValues = newValues;
+            }
+            retVals.put(key, keyValues);
+        }
+
+        return retVals;
     }
 }