+++ sdk/tables/azure-data-tables/src/main/java/module-info.java
@@ -7,6 +7,7 @@ module com.azure.data.tables {
     // public API surface area
     exports com.azure.data.tables;
     exports com.azure.data.tables.models;
+    exports com.azure.data.tables.sas to com.azure.core;
 
     exports com.azure.data.tables.implementation to com.azure.core;
     exports com.azure.data.tables.implementation.models to com.azure.core;
@@ -16,4 +17,5 @@ module com.azure.data.tables {
     opens com.azure.data.tables.implementation to com.fasterxml.jackson.databind, com.azure.core;
     opens com.azure.data.tables.implementation.models to com.fasterxml.jackson.databind, com.azure.core;
     opens com.azure.data.tables.models to com.fasterxml.jackson.databind, com.azure.core;
+    opens com.azure.data.tables.sas to com.fasterxml.jackson.databind, com.azure.core;
 }